+++ src/test/perf/setup.js
@@ -1,0 +1,3 @@
+if (typeof gc !== 'undefined') {
+  beforeBatch(gc);
+}
