+++ toofast.json
@@ -1,0 +1,11 @@
+{
+  "testOptions": {
+    "warmupIterationCount": 100
+  },
+  "setupFiles": [
+    "./src/test/setup.js"
+  ],
+  "include": [
+    "./src/test/**/*.perf.js"
+  ]
+}
